(PC-32696)[PRO] refactor: mv OFFER_LOCATION and SUBCATEGORIES_FIELDS_DEFAULT_VALUES in commons/constants since they are used by both new/old versions of informations screen
