+++ pro/src/pages/IndividualOffer/commons/constants.ts
@@ -1,0 +1,3 @@
+export enum OFFER_LOCATION {
+  OTHER_ADDRESS = 'other',
+}
