+++ pro/src/pages/IndividualOffer/commons/serializers/serializers.spec.ts
@@ -3,8 +3,8 @@
 import { PatchOfferBodyModel, PostOfferBodyModel } from 'apiClient/v1'
 import { AccessibilityEnum } from 'commons/core/shared/types'
 import { getIndividualOfferFactory } from 'commons/utils/individualApiFactories'
+import { OFFER_LOCATION } from 'pages/IndividualOffer/commons/constants'
 import { IndividualOfferFormValues } from 'pages/IndividualOffer/commons/types'
-import { OFFER_LOCATION } from 'pages/IndividualOffer/IndividualOfferDetailsAndInformations/components/OfferLocation/constants'
 
 import {
   serializeDurationMinutes,
